+++ tests/unit/client/feedback/integrations/test_textdescriptives_.py
@@ -16,9 +16,8 @@ from unittest.mock import MagicMock
 
 import pandas as pd
 import pytest
-from argilla import FeedbackDataset
 from argilla.client.feedback.dataset import FeedbackDataset
-from argilla.client.feedback.integrations.textdescriptives_ import TextDescriptivesExtractor
+from argilla.client.feedback.integrations.textdescriptives import TextDescriptivesExtractor
 from argilla.client.feedback.schemas.fields import TextField
 from argilla.client.feedback.schemas.metadata import (
     FloatMetadataProperty,
